+++ NimbleGraphics/NimbleGraphics/include/TextureArray.h
@@ -3,7 +3,6 @@
 
 #include "Typedefs.h"
 #include <vector>
-#include <string>
 #include "Structs.h"
 using std::vector;
 using std::wstring;
@@ -16,14 +15,14 @@ public:
 	~TextureArray();
 
 	ID3D11ShaderResourceView** GetTextureViews();
-	unsigned int GetNumberOfTextures();
+	size_t GetNumberOfTextures();
 	ID3D11ShaderResourceView* GetShaderResourceView(unsigned int index);
 	ID3D11Resource* GetTexture(unsigned int index);
 private:
 	void Create(D3DDevice device, D3DDeviceContext deviceContext, const vector<wstring>& texture_paths);
 	void CreateForSplat(D3DDevice device, D3DDeviceContext deviceContext, const vector<wstring>& texture_paths);
 
-	unsigned int _numTextures;
+	size_t _numTextures;
 	ID3D11ShaderResourceView* _textureViews[MAX_TEXTURES];
 	ID3D11Resource* _textures[MAX_TEXTURES];
 	SplatBuffer _splatBuffer;
